requirements.txt
SQLAlchemy==2.0.43
alembic==1.16.5
apache-airflow==3.1.0
pandas==2.3.2
pendulum==3.1.0
pydantic-settings==2.10.1
pydantic==2.11.9
pymysql==1.1.2
python-dotenv==1.1.1
requests==2.32.5